ES-1837: Add cover for missing compose file, and plugin readme update

+++ tools/corda-runtime-gradle-plugin/src/main/kotlin/net/corda/gradle/plugin/cordalifecycle/CordaLifecycleHelper.kt
@@ -13,6 +13,10 @@ class CordaLifecycleHelper {
         composeFilePath: String,
         dockerProjectName: String
     ) : Process {
+        if (!File(composeFilePath).exists()) {
+            throw CordaRuntimeGradlePluginException("Unable to locate compose file: $composeFilePath")
+        }
+
         val pidStore = PrintStream(FileOutputStream(File(pidFilePath)))
         val cordaProcessBuilder = ProcessBuilder(
             "docker",
